visual-search-project/requirements.txt
# requirements.txt - Python dependencies for Visual Search API

# Core frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML/AI libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
chromadb==0.4.18
Pillow==10.1.0

# Data processing
numpy==1.24.3
pandas==2.1.3

# API utilities
pydantic==2.5.0
python-dotenv==1.0.0

# Async support
aiofiles==23.2.1

# Logging and monitoring
python-json-logger==2.0.7

# CORS support (included in FastAPI)
# Testing (optional)
pytest==7.4.3
httpx==0.25.2

#jwt

PyJWT passlib[bcrypt]